Removing unused imports and casts
git-svn-id: https://svn.apache.org/repos/asf/cocoon/trunk@25815 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/cocoon/components/xpointer/parser/XPointerFrameworkParserTokenManager.java
@@ -314,7 +314,7 @@ public class XPointerFrameworkParserTokenManager
                     }
                 } while (i != startsAt);
             } else {
-                int hiByte = (int) (curChar >> 8);
+                int hiByte = (curChar >> 8);
                 int i1 = hiByte >> 6;
                 long l1 = 1L << (hiByte & 077);
                 int i2 = (curChar & 0xff) >> 6;
@@ -463,7 +463,7 @@ public class XPointerFrameworkParserTokenManager
                     }
                 } while (i != startsAt);
             } else {
-                int hiByte = (int) (curChar >> 8);
+                int hiByte = (curChar >> 8);
                 int i1 = hiByte >> 6;
                 long l1 = 1L << (hiByte & 077);
                 int i2 = (curChar & 0xff) >> 6;
